# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /modules:
    get:
      operationId: getModules
      tags: [modules]
      summary: Get modules for a course
      parameters:
        - name: courseId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleRequest"
      responses:
        "201":
          description: Created module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{id}:
    put:
      operationId: updateModule
      tags: [modules]
      summary: Update module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModuleRequest"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete module
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /lessons:
    get:
      operationId: getLessons
      tags: [lessons]
      summary: Get lessons for a module
      parameters:
        - name: moduleId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Created lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLessonById
      tags: [lessons]
      summary: Get lesson by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson with quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonDetail"
    put:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonRequest"
      responses:
        "200":
          description: Updated lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /quizzes:
    get:
      operationId: getQuizByLesson
      tags: [quizzes]
      summary: Get quiz for a lesson
      parameters:
        - name: lessonId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create quiz for a lesson
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuizRequest"
      responses:
        "201":
          description: Created quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{id}:
    put:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuizRequest"
      responses:
        "200":
          description: Updated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete quiz
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Module:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        orderIndex:
          type: integer
        lessonCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, courseId, title, orderIndex, lessonCount, createdAt, updatedAt]
    CreateModuleRequest:
      type: object
      properties:
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        orderIndex:
          type: integer
      required: [courseId, title]
    UpdateModuleRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        orderIndex:
          type: integer
...
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        orderIndex:
          type: integer
        quiz:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/QuizSummary"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, moduleId, title, orderIndex, createdAt, updatedAt]
    LessonDetail:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        orderIndex:
          type: integer
        quiz:
          nullable: true
          allOf:
            - $ref: "#/components/schemas/Quiz"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, moduleId, title, orderIndex, createdAt, updatedAt]
    CreateLessonRequest:
      type: object
      properties:
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        orderIndex:
          type: integer
      required: [moduleId, title]
    UpdateLessonRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        duration:
          type: integer
          nullable: true
        orderIndex:
          type: integer
    QuizQuestion:
      type: object
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        answer:
          type: string
      required: [question, options, answer]
    QuizSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        lessonId:
          type: integer
      required: [id, title, lessonId]
    Quiz:
      type: object
      properties:
        id:
          type: integer
        lessonId:
          type: integer
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, lessonId, title, questions, createdAt, updatedAt]
    CreateQuizRequest:
      type: object
      properties:
        lessonId:
          type: integer
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
      required: [lessonId, questions]
    UpdateQuizRequest:
      type: object
      properties:
        title:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
...