# lib/api-zod/src/generated/types/quizSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizSummary {
  id: number;
  title: string;
  lessonId: number;
}